addon.d Script Implementation
This commit implements a addon.d script for compatible kernels that will backup the
v4a driver and updated audio conf files so they survive the flashing of a rom update
where supported.

The script will be automatically installed on supported devices during the installation of
the v4a driver.