SuperUser Compatibility Fix
This commit provides a work-around using existing methods to check for the presence
of Chainfire's SuperSU and if this is not found executes the root commands in a different
manner to take into account issues with SuperUser.

I have only implemented this within the VBOX installation method as this is where
I would experience all failed installs with SuperUser when debugging, if the RootTools
installation method also causes issues then the InstallDrv_FX method could be tweaked
to call the VBOX method if SuperSU is not installed thus using the modifed su command
structure.

Tested with both SuperUser & SuperSU with both functioning as expected, hopefully
this once and for all removes the issues faced with installing the driver using
SuperUser making the installation more user friendly and stopping about 90% of
posts on the xda thread which are all about the driver failing to install!! :)