Driver Installation Re-write
This commit proposes a switch back to a RootTools/Busybox installation method without
including the recently added vBox methods and library.

Busybox detection has been incorporated into the driver installation method allowing
for the user to be provided the option to install Busybox through the Playstore.

Driver installation in proceed through RootTools method first then if required will fall back
onto a direct execution style using the RootExecuteWithoutShell method.

The RootTools library was updated to version 3.4 with the updated version addressing
bugs and issues found in the previous version used 2.6.

***** Build.prop lpa Detection Added *****

Build.prop Tools v1.1.5 beta has been added to the project and a new method BuildPropOk added which checks for the lpa properties which can cause issues with v4a, these are:
    Ipa.decode=true to lpa.decode=false
    tunnel.decode=true to tunnel.decode=false
    lpa.use-stagefright=true to lpa.use-stagefright=false

If any of the properties are located the user is prompted to update the required entries
with a warning about backing up the system in case of failure.